refactor: to test_gradients.py
