+++ tests/test_gradients.py
@@ -1,10 +1,13 @@
 from typing import List
 
+import numpy as np
 import pytest
 import torch
 
-from pytorch_optimizer import load_optimizer
+from pytorch_optimizer import SAM, AdamP, load_optimizer
 from pytorch_optimizer.base.exception import NoSparseGradientError
+from tests.test_optimizers import OPTIMIZERS
+from tests.utils import build_environment, ids, tensor_to_numpy
 
 SPARSE_OPTIMIZERS: List[str] = [
     'madgrad',
@@ -30,6 +33,52 @@ NO_SPARSE_OPTIMIZERS: List[str] = [
     'adapnm',
     'pnm',
 ]
+
+
+def test_sam_no_gradient():
+    (x_data, y_data), model, loss_fn = build_environment()
+    model.fc1.weight.requires_grad = False
+    model.fc1.weight.grad = None
+
+    optimizer = SAM(model.parameters(), AdamP)
+    optimizer.zero_grad()
+
+    loss = loss_fn(y_data, model(x_data))
+    loss.backward()
+    optimizer.first_step(zero_grad=True)
+
+    loss_fn(y_data, model(x_data)).backward()
+    optimizer.second_step(zero_grad=True)
+
+
+@pytest.mark.parametrize('optimizer_config', OPTIMIZERS, ids=ids)
+def test_no_gradients(optimizer_config):
+    (x_data, y_data), model, loss_fn = build_environment()
+
+    model.fc1.weight.requires_grad = False
+    model.fc1.bias.requires_grad = False
+
+    optimizer_class, config, iterations = optimizer_config
+    optimizer = optimizer_class(model.parameters(), **config)
+
+    if optimizer_class.__name__ == 'Nero':
+        pytest.skip(f'skip {optimizer_class.__name__}')
+
+    init_loss, loss = np.inf, np.inf
+    for _ in range(iterations):
+        optimizer.zero_grad()
+
+        y_pred = model(x_data)
+        loss = loss_fn(y_pred, y_data)
+
+        if init_loss == np.inf:
+            init_loss = loss
+
+        loss.backward()
+
+        optimizer.step()
+
+    assert tensor_to_numpy(init_loss) >= tensor_to_numpy(loss)
 
 
 @pytest.mark.parametrize('no_sparse_optimizer', NO_SPARSE_OPTIMIZERS)
